♻️ Simplify and unify compat layer
